refactor(docsite): extract migration validation helpers
